+++ app/src/main/java/com/techcourse/DispatcherServlet.java
@@ -3,17 +3,12 @@ package com.techcourse;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
 import com.interface21.webmvc.servlet.mvc.HandlerAdaptor;
-import com.interface21.webmvc.servlet.mvc.controller.ControllerAdaptor;
-import com.interface21.webmvc.servlet.mvc.handler.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.handler.HandlerExecutionAdaptor;
-import com.interface21.webmvc.servlet.mvc.handler.HandlerMapping;
+import com.interface21.webmvc.servlet.mvc.HandlerAdaptors;
+import com.interface21.webmvc.servlet.mvc.HandlerMappings;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -22,27 +17,16 @@ public class DispatcherServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
 
-    private final List<HandlerMapping> handlerMappings = new ArrayList<>();
-    private final List<HandlerAdaptor> handlerAdaptors = new ArrayList<>();
+    private final HandlerMappings handlerMappings;
+    private final HandlerAdaptors handlerAdaptors;
 
     public DispatcherServlet() {
+        this.handlerMappings = new HandlerMappings("com.techcourse.controller");
+        this.handlerAdaptors = new HandlerAdaptors();
     }
 
     @Override
     public void init() {
-        HandlerMapping manualHandlerMapping = new ManualHandlerMapping();
-        manualHandlerMapping.initialize();
-        handlerMappings.add(manualHandlerMapping);
-
-        HandlerMapping annotationHandlerMapping = new AnnotationHandlerMapping("com.techcourse.controller");
-        annotationHandlerMapping.initialize();
-        handlerMappings.add(annotationHandlerMapping);
-
-        ControllerAdaptor controllerAdaptor = new ControllerAdaptor();
-        handlerAdaptors.add(controllerAdaptor);
-
-        HandlerExecutionAdaptor handlerExecutionAdaptor = new HandlerExecutionAdaptor();
-        handlerAdaptors.add(handlerExecutionAdaptor);
     }
 
     @Override
@@ -52,16 +36,8 @@ public class DispatcherServlet extends HttpServlet {
         log.debug("Method : {}, Request URI : {}", request.getMethod(), requestURI);
 
         try {
-            Object handler = handlerMappings.stream()
-                    .map(handlerMapping -> handlerMapping.getHandler(request))
-                    .filter(Objects::nonNull)
-                    .findAny()
-                    .orElseThrow();
-
-            HandlerAdaptor adaptor = handlerAdaptors.stream()
-                    .filter(handlerAdaptor -> handlerAdaptor.canExecute(handler))
-                    .findAny()
-                    .orElseThrow();
+            Object handler = handlerMappings.getHandler(request);
+            HandlerAdaptor adaptor = handlerAdaptors.findAdaptor(handler);
 
             ModelAndView modelAndView = adaptor.execute(handler, request, response);
