+++ mvc/src/main/java/com/interface21/webmvc/servlet/HandlerExecutionAdaptor.java
@@ -1,7 +1,7 @@
-package com.interface21.webmvc.servlet.mvc.handler;
+package com.interface21.webmvc.servlet;
 
-import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.mvc.HandlerAdaptor;
+import com.interface21.webmvc.servlet.mvc.handler.HandlerExecution;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
