+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/HandlerMappings.java
@@ -1,0 +1,31 @@
+package com.interface21.webmvc.servlet.mvc;
+
+import com.interface21.webmvc.servlet.mvc.handler.AnnotationHandlerMapping;
+import com.techcourse.ManualHandlerMapping;
+import jakarta.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+public class HandlerMappings {
+
+    private final List<HandlerMapping> handlerMappings = new ArrayList<>();
+
+    public HandlerMappings(final Object... basePackage) {
+        HandlerMapping manualHandlerMapping = new ManualHandlerMapping();
+        manualHandlerMapping.initialize();
+        handlerMappings.add(manualHandlerMapping);
+
+        HandlerMapping annotationHandlerMapping = new AnnotationHandlerMapping(basePackage);
+        annotationHandlerMapping.initialize();
+        handlerMappings.add(annotationHandlerMapping);
+    }
+
+    public Object getHandler(final HttpServletRequest request) {
+        return handlerMappings.stream()
+                .map(handlerMapping -> handlerMapping.getHandler(request))
+                .filter(Objects::nonNull)
+                .findAny()
+                .orElseThrow();
+    }
+}
